css_fixed: Avoid undefined shift.
From undefined behaviour sanitizer:

  src/utils/utils.c:130:18: runtime error: left shift of negative value -1

+++ src/utils/utils.c
@@ -127,6 +127,6 @@ css_fixed css__number_from_string(const uint8_t *data, size_t len,
 		}
 	}
 
-	return (intpart << 10) | fracpart;
+	return ((uint32_t)intpart << 10) | fracpart;
 }
